More Changes. There were not as many as last time.
Agenda: More specific commits.

+++ src/net/pvschools/robotics/javabot/practice/commands/RoboCommand.java
@@ -26,7 +26,7 @@ public abstract class RoboCommand extends Command
         super();
     }
 
-    public RoboCommand(int timeOut)
+    public RoboCommand(double timeOut)
     {
         super(timeOut);
     }
